More compiler precision warnings

+++ ibuffer suite/ibufconcatenate~/ibufconcatedrive~.cpp
@@ -111,7 +111,7 @@ void ibufconcatedrive_free(t_ibufconcatedrive *x)
 void store(float *lo_res, float *hi_res, double output)
 {
     *lo_res = static_cast<float>(output);
-    *hi_res = output - *lo_res;
+    *hi_res = static_cast<float>(output - *lo_res);
 }
 
 void store(double *lo_res, double *hi_res, double output)
